+++ packages/in_app_purchase/in_app_purchase_storekit/pigeons/sk2_pigeon.dart
@@ -261,6 +261,9 @@ abstract class InAppPurchase2API {
 
   @async
   void presentOfferCodeRedeemSheet();
+
+  @async
+  SK2TransactionMessage? latestTransaction(String productId);
 }
 
 @FlutterApi()